requirements.txt
djangorestframework
Django~=4.2.3